Fix file references
Old broken build-css script in package.json:

"scripts": {
  "build-css": "node-sass --include-path scss src/App.scss src/App.css",
}

Previously, the reference to the non-existing "src/App.scss" produces
a not-appropriate error of:

"An output directory must be specified when compiling a directory"

"src/App.css" is the correct specified output directory, but the
reference to a non-existing "src/App.scss" file somehow botches up
the parsing of the output directory.

New build-css script in package.json:

"scripts": {
  "build-css": "node-sass --include-path src/App.scss src/App.sass src/App.css",
}

As a new fix, I included both "src/App.scss" and "src/App.sass" as
inputs so that we can use both dialects of SASS to get compiled down
to "src/App.css".

The SASS dialect is like Ruby with no braces and semicolons. The SCSS
dialect looks very close to standard CSS and uses curly braces and
semicolons.