Fix build-css script
The old build script causes this error:
"An output directory must be specified when compiling a directory"

For some reason, the "--include-path" directive is very finicky. Here is
the new script:

"scripts": {
  "build-css": "node-sass --include-path src src/App.sass src/App.css"
}

It seems like you need all 3 parameters: a "src" for the starting
directory, then the reference to the .sass file, and finally the output
file of "src/App.css".

For simplicity, I'm assuming that we will all use the SASS syntax and
ignore adapting the build-css script to handle the SCSS syntax.